+++ src/SparseMatrixConverter.cpp
@@ -7,7 +7,7 @@ SparseMatrixConverter::SparseMatrixConverter() {
 
 }
 
-void SparseMatrixConverter::fullToPodArray(const Array<double, 2> & A, int * Apod) {
+void SparseMatrixConverter::fullToPodArray(const Array<double, 2> & A, double * Apod) {
     int ind = 0;
     for (int i =0; i< A.rows(); i++) {
         for (int j=0; j< A.cols(); j++) {
